Updating dependencies and doing a sync
i updated the dependencies that are necessary for my part.
I tried to keep everything as clean and updated as possible, and the sync was a success.

The dependencies for cardview and design, that I need for testing were throwing me an error so I will have to face that problem later