success!
I made a translation of LiveData List<Note> to ArrayList<Note>. It got me all day to achieve that. For you it means that you will have the data inside an arrayList and ready to use. With LiveData you will had to use Observers all the time, and direct access to the content is was almost impossible.
Now you can work with both formats.
Now I will transfer the functionality to the actula lists and everything will be set to go.
I will do soon an advance testing to see if the updating works as it should.