It freaking works.
It works.
Of course I will have to achieve better control of the DataBase because the thing that I have now, is a quick fix, and I will have of course smooth some rough edges and do a better testing. that includes updating an object.
Comments and other details need to be fixed as well, but the thing is that it works.