Tranfer seems to be succesful
it seems that the tranfer of the data from the DataHolder to the live data is succesful. the problem is that the data takes time to load and that gives sometimes that makes the app to crash the first time.
The problem lies to the async. and it is nothing that bad.