pullFromDataHolderToTheArrayLists method
this method pulls the data from the ArrayLists of DataHolder to the ArrayLists of the ViewModel.
This method can't be called from the DataHolder and it needs to be called from an activity that has ViewModel or the NotViewModel itself.
This method is used for synch and usually comes before the syncLiveDataLists() method.