if idExists check
this little check ensures that will not be doublecreated users and also that the user with id 0 will not be updated if the id is wrong.
The best security of course is to let the data load before you do any manupulation and to avoid to do automatic  data manipulation in the instantiation of an activity.
Otherwise my solution is almost all in one solution.