adding a default image for the items
when testing or when there is not image specified for an item, there should be a default image. For now I selected that whih can be change later.
There is agood possibility that for accessing the image we should set the default value as "photoname.jpg" or something else, but this has to be seen.