reset password is debugged.
The issue was that the new password was not set with encyption. Now the issue is solved.
Now I will have to figure out why the check for the email doesnt work.
It is trivial and it will probably removed as it doesn't help on anything but i will try to fix it anyway.