Generation of the order is complete.
Now the shopping cart produces a nice order (Note Object).
Before I call it a day, I will add the option to make the payment in the same activity,
It will be an option controlled by a boolean.
If true the paymen will be done here and the next activity will be the invoice.
If false, the payment will be done in the next activity