debugging: Gmailsender and checkout
The code of the gmailsender is legit. For not giving a legit  eception. I assume that the problem lies in the libraries and that exactly will be my attack plan.
Also the application now does not a checkout if no items are in the shopping cart.

I will make a break before I start optimizing jars.